+++ src/main/java/kitchenpos/domain/OrderLineItem.java
@@ -10,6 +10,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
+import kitchenpos.menu.domain.Menu;
 
 import java.math.BigDecimal;
 import java.util.UUID;
